fixed cloudinary

+++ src/services/profile.service.ts
@@ -1,5 +1,6 @@
 import type { IProfileRepository } from "../repository/profile.repository";
 import type { Profile } from "../generated";
+import cloudinary from "../utils/cloudinary";
 
 export interface IProfileService {
   create(data: {
@@ -7,6 +8,7 @@ export interface IProfileService {
     gender: string;
     address: string;
     profile_picture_url?: string;
+    public_id?: string;
     userId: number;
   }): Promise<Profile>;
 
@@ -18,19 +20,21 @@ export interface IProfileService {
     gender: string;
     address: string;
     profile_picture_url?: string;
+    public_id?: string;
   }>): Promise<Profile>;
 
   delete(id: number): Promise<void>;
 }
 
 export class ProfileService implements IProfileService {
-  constructor(private profileRepo: IProfileRepository) {}
+  constructor(private profileRepo: IProfileRepository) { }
 
   async create(data: {
     name: string;
     gender: string;
     address: string;
     profile_picture_url?: string;
+    public_id?: string;
     userId: number;
   }): Promise<Profile> {
     const exist = await this.profileRepo.findByUserId(data.userId);
@@ -58,14 +62,26 @@ export class ProfileService implements IProfileService {
       gender: string;
       address: string;
       profile_picture_url?: string;
+      public_id?: string;
     }>
   ): Promise<Profile> {
-    await this.getById(id); // pastiin ada
+    const profile = await this.profileRepo.findByUserId(id);
+    if (!profile) throw new Error("Profile tidak ditemukan");
+
+    if (data.public_id && profile.public_id) {
+      await cloudinary.uploader.destroy(profile.public_id);
+    }
+
     return this.profileRepo.update(id, data);
   }
 
   async delete(id: number): Promise<void> {
-    await this.getById(id); // pastiin ada
+    const profile = await this.profileRepo.findByUserId(id);
+    if (!profile) throw new Error("Profile tidak ditemukan");
+
+    if (profile.public_id) {
+      await cloudinary.uploader.destroy(profile.public_id);
+    }
     await this.profileRepo.delete(id);
   }
 }
